Add files via upload
Roshambo is an alternative name for the guessing game known as Scissors, Rock, and Paper.  Two players pick one of the three simultaneously.  The goal is to pick the one that beats the opponent's pick.  Scissors beats Paper.  Paper beats Rock.  Rock beats Scissors.  
The Roshambo script, written in powershell, provides a basic, character oriented interface.  There is an adaptive strategy built in,  but it isn't that hard to beat.